Refactor: change folder structure
